requirements.txt
datasets
huggingface-hub
ipywidgets
matplotlib
numpy
opencv-python
Pillow
plotly
pycolmap
scipy
torch
trimesh
webdataset
manifold3d # for metric computation